backend/requirements.txt
langchain
langchain_chroma
langchain_openai
langchain-openai
click
fastapi[standard]
pyjwt
passlib
bcrypt==4.0.1
sqlmodel
PyMySQL
google-cloud-vision
pillow
protobuf==3.20.0
cryptography